A few small (but noisy) changes! * Rename NMOS_HOME_DIR to NMOS_CPP_DIR and NmosCppGeneral to NmosCppCommon * Rename nmos::experimental::node_resources_thread to node_implementation_thread * Rename nmos-cpp-registry-test to nmos-cpp-test (another long-overdue change, since it includes unit tests relevant to both registry and node) * Fix a couple of typos, delete some whitespace and commented out lines
